refactor: consolidate VM manager to single cross-platform implementation
This refactors the VM manager from platform-specific files (vm_linux.go, vm_darwin.go)
to a single cross-platform implementation (manager.go) with minimal platform-specific
backend factories.

Changes:
- internal/vm/manager.go: New cross-platform VM manager with all shared logic
- internal/vm/platform_linux.go: Linux-specific backend factory
- internal/vm/platform_darwin.go: macOS-specific backend factory
- cmd/tent/create.go: Updated to use platform-specific backend factory
- Removed: internal/vm/vm_linux.go, internal/vm/vm_darwin.go (code duplication)

Benefits:
- Reduced code duplication (~25% line reduction)
- Single source of truth for VM lifecycle logic
- Easier to maintain and test

All tests pass, build succeeds, go vet clean.
