+++ project/coverage.out
@@ -437,27 +437,6 @@ github.com/dalinkstone/tent/internal/storage/storage.go:377.32,378.64 1 1
 github.com/dalinkstone/tent/internal/storage/storage.go:378.64,380.18 2 1
 github.com/dalinkstone/tent/internal/storage/storage.go:380.18,386.5 1 1
 github.com/dalinkstone/tent/internal/storage/storage.go:390.2,390.20 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:9.38,11.2 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:18.42,19.24 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:19.24,21.3 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:22.2,22.28 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:26.59,28.2 1 1
-github.com/dalinkstone/tent/pkg/models/errors.go:31.44,33.2 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:14.35,16.2 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:88.44,89.16 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:89.16,93.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:95.2,97.20 2 1
-github.com/dalinkstone/tent/pkg/models/types.go:97.20,99.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:101.2,101.20 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:101.20,103.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:105.2,105.23 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:105.23,107.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:109.2,109.21 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:109.21,111.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:113.2,113.21 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:113.21,115.3 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:117.2,117.12 1 1
-github.com/dalinkstone/tent/pkg/models/types.go:121.39,123.2 1 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:69.156,70.25 1 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:70.25,72.17 2 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:72.17,74.4 1 0
@@ -561,3 +540,24 @@ github.com/dalinkstone/tent/internal/vm/vm_linux.go:397.92,400.54 2 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:400.54,402.55 2 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:402.55,404.4 1 1
 github.com/dalinkstone/tent/internal/vm/vm_linux.go:408.2,412.8 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:9.38,11.2 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:18.42,19.24 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:19.24,21.3 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:22.2,22.28 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:26.59,28.2 1 1
+github.com/dalinkstone/tent/pkg/models/errors.go:31.44,33.2 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:14.35,16.2 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:88.44,89.16 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:89.16,93.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:95.2,97.20 2 1
+github.com/dalinkstone/tent/pkg/models/types.go:97.20,99.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:101.2,101.20 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:101.20,103.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:105.2,105.23 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:105.23,107.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:109.2,109.21 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:109.21,111.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:113.2,113.21 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:113.21,115.3 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:117.2,117.12 1 1
+github.com/dalinkstone/tent/pkg/models/types.go:121.39,123.2 1 1
