+++ project/iteration_27_summary.json
@@ -1,0 +1,11 @@
+{
+  "iteration": 27,
+  "action_type": "fix",
+  "summary": "Implemented cross-platform VM manager by splitting vm.go into vm_linux.go and vm_darwin.go with build tags",
+  "result": "success",
+  "pr_number": 31,
+  "merged": true,
+  "confidence": 0.75,
+  "hypothesis_result": "confirmed",
+  "error": null
+}
